# apps/api_v2/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Coder API v2 - User, AI, Challenge, TestCases, Courses, Exams, ExamItems, Submissions and Sessions Endpoints
  version: 1.0.0
  description: |
    OpenAPI contract built from the real HTTP handlers in apps/api_v2/internal/interfaces/http.

    Covered endpoints:
    - POST /auth/register
    - POST /auth/login
    - POST /auth/refresh-token
    - GET /auth/me
    - POST /ai/generate-challenge-ideas
    - POST /ai/generate-test-cases
    - POST /ai/generate-full-challenge
    - POST /ai/generate-exam
    - POST /challenges/
    - GET /challenges/
    - GET /challenges/public
    - GET /challenges/{id}
    - PATCH /challenges/{id}
    - DELETE /challenges/{id}
    - POST /challenges/{id}/publish
    - POST /challenges/{id}/fork
    - POST /challenges/{id}/archive
    - POST /test-cases/
    - GET /test-cases/challenge/{challengeId}
    - PATCH /test-cases/{id}
    - DELETE /test-cases/{id}
    - POST /courses/enroll
    - POST /courses/
    - GET /courses/
    - GET /courses/{id}
    - POST /courses/{id}
    - DELETE /courses/{id}
    - POST /courses/{id}/students
    - GET /courses/{id}/students
    - DELETE /courses/{id}/students/{studentId}
    - POST /exams/
    - POST /exams/default-code-templates
    - GET /exams/public
    - GET /exams/course/{courseId}
    - GET /exams/{id}
    - PATCH /exams/{id}
    - DELETE /exams/{id}
    - POST /exams/{id}/visibility
    - POST /exams/{id}/close
    - GET /exams/{examId}/items
    - POST /exam-items/
...

...
servers:
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      description: Creates a new user account and returns token plus basic user data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              student_register:
                value:
                  email: student@example.com
                  name: Student Test
                  password: Password123!
      responses:
        '201':
          description: User created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccess'
              examples:
                created:
                  value:
                    token:
                      access_token: jwt-access-token
                      refresh_token: jwt-refresh-token
                    user_data:
                      id: usr-123
                      username: Student Test
                      email: student@example.com
                      role: student
        '400':
          description: Invalid payload or domain validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                bad_request:
                  value:
                    error: invalid input data
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      description: Validates credentials and returns access plus refresh tokens and user data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              teacher_login:
                value:
                  email: test@test.com
                  password: Password123!
      responses:
        '200':
          description: Authenticated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAccess'
              examples:
                ok:
                  value:
                    token:
                      access_token: jwt-access-token
                      refresh_token: jwt-refresh-token
                    user_data:
                      id: usr-123
                      username: Student Test
                      email: student@example.com
                      role: student
        '400':
          description: Invalid credentials or bad payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                bad_credentials:
                  value:
                    error: invalid credentials
  /auth/refresh-token:
    post:
      tags: [Auth]
      summary: Refresh auth token
      description: Issues a new access token from a valid refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
            examples:
              refresh_request:
                value:
                  refresh_token: jwt-refresh-token
      responses:
        '200':
          description: Token refreshed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
              examples:
                refreshed:
                  value:
                    access_token: new-access-token
                    refresh_token: new-refresh-token
        '400':
          description: Invalid refresh token or malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_refresh:
                  value:
                    error: invalid refresh token
  /auth/me:
    get:
      tags: [Auth]
      summary: Get user profile
      description: |
        Returns the user profile from UserModule.GetData.

        Real handler behavior:
        - Resolves user identity from query param `userId` first.
        - If `userId` is empty, falls back to `X-User-Email` header.
        - If both are missing, returns 400 with:
          `{"error": "userId or X-User-Email header is required"}`.
      parameters:
        - in: query
          name: userId
          required: false
          schema:
            type: string
          description: User identifier. If omitted, handler uses header X-User-Email as fallback lookup key.
        - in: header
          name: X-User-Email
          required: false
          schema:
            type: string
            format: email
          description: Fallback user lookup value when userId is not provided.
        - in: header
          name: Authorization
          required: false
          schema:
            type: string
          description: Optional Bearer token propagated in request context.
      responses:
        '200':
          description: User profile returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples:
                student:
                  value:
                    id: usr-123
                    username: Student Test
                    email: student@example.com
                    role: student
        '400':
          description: Missing user selector or domain error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_selector:
                  value:
                    error: userId or X-User-Email header is required
  /ai/generate-full-challenge:
    post:
      tags: [AI]
      summary: Generate full programming challenge with AI
      description: |
        Uses AIModule.GenerateFullChallenge and returns a generated challenge structure
        including statement, variables and suggested public/hidden test cases.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateFullChallengeRequest'
            examples:
              medium_array:
                value:
                  topic: arrays
                  difficulty: medium
      responses:
        '200':
          description: Challenge generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateFullChallengeOutput'
              examples:
                generated:
                  value:
                    challenge:
                      title: Suma de pares en arreglo
                      description: Dado un arreglo de enteros, calcula la suma de los elementos pares.
                      difficulty: medium
                      tags: [arrays, loops]
                      worker_time_limit: 1500
                      worker_memory_limit: 256
                      input_variables:
                        - name: values
                          type: array
                          value: "[1,2,3,4]"
                      output_variable:
                        name: sum_even
                        type: int
                        value: "6"
                      constraints: 1 <= n <= 100000
                      public_test_cases:
                        - name: Ejemplo 1
                          visible: true
                          input:
                            - name: values
                              type: array
                              value: "[1,2,3,4]"
                          output:
                            name: sum_even
                            type: int
                            value: "6"
                      hidden_test_cases:
                        - name: Oculto 1
                          visible: false
                          input:
                            - name: values
                              type: array
                              value: "[100,101,102]"
                          output:
                            name: sum_even
                            type: int
                            value: "202"
        '400':
          description: Invalid payload or generation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ai/generate-exam:
    post:
      tags: [AI]
      summary: Generate exam with AI
      description: Uses AIModule.GenerateExam and returns generated exam title, description and limits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateExamRequest'
            examples:
              easy_oop:
                value:
                  topic: Programacion orientada a objetos
                  difficulty: easy
      responses:
        '200':
          description: Exam generated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateExamOutput'
              examples:
                generated:
                  value:
                    exam:
                      title: Examen introductorio de POO
                      description: Evalua conceptos basicos de clases, objetos y encapsulamiento.
                      time_limit: 90
                      try_limit: 2
        '400':
          description: Invalid payload or generation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /challenges/:
    post:
      tags: [Challenges]
      summary: Create challenge
      description: Creates a challenge in the authenticated teacher repository.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeCreateRequest'
            examples:
              create:
                value:
                  title: Suma de dos numeros
                  description: Dado a y b, retorna la suma.
                  tags: [math, basics]
                  status: draft
                  difficulty: easy
                  worker_time_limit: 1500
                  worker_memory_limit: 256
                  code_templates:
                    python: "def solve():\n    pass"
                  input_variables:
                    - name: a
                      type: int
                      value: "2"
                    - name: b
                      type: int
                      value: "3"
                  output_variable:
                    name: result
                    type: int
                    value: "5"
                  constraints: "-10^9 <= a,b <= 10^9"
      responses:
        '201':
          description: Challenge created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeResponse'
              examples:
                created:
                  value:
                    title: Suma de dos numeros
                    description: Dado a y b, retorna la suma.
                    tags:
                      - math
                      - basics
                    status: draft
                    difficulty: easy
                    worker_time_limit: 1500
                    worker_memory_limit: 256
                    code_templates:
                      python: "def solve():\n    pass"
                    input_variables:
                      - name: a
                        type: int
                        value: "2"
                      - name: b
                        type: int
                        value: "3"
                    output_variable:
                      name: result
                      type: int
                      value: "5"
                    constraints: "-10^9 <= a,b <= 10^9"
        '400':
          description: Invalid payload or domain validation error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Challenges]
      summary: List user challenges
      description: Returns challenges visible to the authenticated user context.
      parameters:
        - in: query
          name: examId
          required: false
          schema:
            type: string
          description: Optional exam identifier filter.
      responses:
        '200':
          description: Challenge list returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChallengeResponse'
              examples:
                list:
                  value:
                    - title: Suma de dos numeros
                      description: Dado a y b, retorna la suma.
                      tags:
                        - math
                        - basics
                      status: draft
                      difficulty: easy
                      worker_time_limit: 1500
                      worker_memory_limit: 256
                      code_templates:
                        python: "def solve():\n    pass"
                      input_variables:
                        - name: a
                          type: int
                          value: "2"
                        - name: b
                          type: int
                          value: "3"
                      output_variable:
                        name: result
                        type: int
                        value: "5"
                      constraints: "-10^9 <= a,b <= 10^9"
        '400':
          description: Query validation or domain error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /challenges/public:
    get:
      tags: [Challenges]
      summary: List public challenges
      description: Returns published/public challenges with optional filters.
      parameters:
        - in: query
          name: tag
          required: false
          schema:
            type: string
        - in: query
          name: difficulty
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Public challenge list returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChallengeResponse'
              examples:
                public_list:
                  value:
                    - title: Suma de dos numeros
                      description: Dado a y b, retorna la suma.
                      tags:
                        - math
                        - basics
                      status: draft
                      difficulty: easy
                      worker_time_limit: 1500
                      worker_memory_limit: 256
                      code_templates:
                        python: "def solve():\n    pass"
                      input_variables:
                        - name: a
                          type: int
                          value: "2"
                        - name: b
                          type: int
                          value: "3"
                      output_variable:
                        name: result
                        type: int
                        value: "5"
                      constraints: "-10^9 <= a,b <= 10^9"
        '400':
          description: Query validation or domain error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /challenges/default-code-templates:
    post:
      tags: [Challenges]
      summary: Generate default code templates by language
      description: |
        Generates starter code templates for all supported languages using input/output variable definitions.
        This endpoint is restricted to professor role in the current use case.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExamDefaultCodeTemplatesRequest'
            examples:
              basic_math:
                value:
                  input_variables:
                    - name: a
                      type: int
                      value: "2"
                    - name: b
                      type: int
                      value: "3"
                  output_variable:
                    name: result
                    type: int
                    value: "5"
      responses:
        '200':
          description: Default templates generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamDefaultCodeTemplatesResponse'
              examples:
                templates:
                  value:
                    python: "def solve(a, b):\n    # TODO\n    return 0"
                    javascript: "function solve(a, b) {\n  // TODO\n  return 0;\n}"
                    java: "public class Main {\n  public static int solve(int a, int b) {\n    // TODO\n    return 0;\n  }\n}"
                    cpp: "int solve(int a, int b) {\n  // TODO\n  return 0;\n}"
        '400':
          description: Invalid payload or domain/permission error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /challenges/{id}/publish:
    post:
      tags: [Challenges]
      summary: Publish challenge
      description: |
        Changes challenge status to published when the authenticated user has permission.
        This handler returns the domain entity directly.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Challenge published.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeEntityResponse'
        '400':
          description: Invalid id or permission/domain error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /challenges/{id}/fork:
    post:
      tags: [Challenges]
      summary: Fork challenge
      description: Creates a fork copy in the authenticated teacher repository.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Challenge forked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeEntityResponse'
        '400':
          description: Invalid id or permission/domain error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        password:
          type: string
          minLength: 1
...